# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is an open API for `BAD` operations management. You can use the\
    \ api key `api_key` to test the authorization filters. <h2>Authentication</h2>\
    \ All API's must be authenticated. To authenticate, supply your API key as the\
    \ `Authorization` HTTP header with the `api_key` authorization scheme. <h2>Parameters</h2>\
    \ For API's that use POST, set the request's Content-Type header to `application/json`\
    \ and supply an appropriate JSON object in the request body. For API's that use\
    \ GET, specify each request parameter as a properly-escaped query parameter. <h2>Response</h2>\
    \ All responses have their Content-Type set to `application/json`, and contain\
    \ the response JSON object in the response body."
  version: "1.0.0"
  title: "BAD and Certified pick-ups OpenAPI"
host: "localhost:8081"
basePath: "/v2"
...
schemes:
- "http"
consumes:
- "application/json"
paths:
...
  /createCPU:
    post:
      tags:
      - "Cpu"
      summary: "Create a new CPU for a given BAD."
      description: "Creates a new CPU for a given BAD in the creation phase. The minter\
        \ indicates the first owner (transporter).\n"
      operationId: "createCPUPOST"
      parameters:
      - in: "body"
        name: "body"
        description: "CPU object"
        required: true
        schema:
          $ref: "#/definitions/CpuWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/CpuResponse"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/CpuResponse"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/CpuResponse"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/CpuResponse"
      x-swagger-router-controller: "Cpu"
  /readCPU:
    get:
      tags:
      - "Cpu"
      summary: "Read a CPU for a given CPUId."
      description: "Reads a CPU for a given CPUId.\n"
      operationId: "readCPUGET"
      parameters:
      - name: "nonce"
        in: "query"
        description: "nonce of the CPU"
        required: true
        type: "string"
      - name: "badId"
        in: "query"
        description: "ID of the bad"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Cpu"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/CpuResponse"
        "404":
          description: "CPU not found"
          schema:
            $ref: "#/definitions/CpuResponse"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/CpuResponse"
      x-swagger-router-controller: "Cpu"
  /readCPUByTokenValue:
    get:
      tags:
      - "Cpu"
      summary: "Read a CPU using its token value."
      description: "Reads a CPU for a given CPUId.\n"
      operationId: "readCPUByTokenValueGET"
      parameters:
      - name: "tokenValue"
        in: "query"
        description: "the random token vlaue in the CPU"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Cpu"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/CpuResponse"
        "404":
          description: "CPU not found"
          schema:
            $ref: "#/definitions/CpuResponse"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/CpuResponse"
      x-swagger-router-controller: "Cpu"
  /reAssignCpu:
    put:
      tags:
      - "Cpu"
      summary: "Assign a CPU to a new owner."
      description: "Assigns a CPU to a new owner in the blockchain.\n"
      operationId: "reAssignCPUPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "CPU object"
        required: true
        schema:
          $ref: "#/definitions/assignCpuRequestWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/CpuResponse"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/CpuResponse"
        "404":
          description: "CPU not found"
          schema:
            $ref: "#/definitions/CpuResponse"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/CpuResponse"
      x-swagger-router-controller: "Cpu"
  /supplementCpu:
    put:
      tags:
      - "Cpu"
      summary: "Supplement a CPU to a driver and transfer ownership to the driver."
      description: "Supplements a CPU to a driver and transfers ownership to the driver\
        \ in the blockchain with the provided details. This function is called by\
        \ the owner of the CPU (transporter).\n"
      operationId: "supplementCPUPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "CPU object"
        required: true
        schema:
          $ref: "#/definitions/supplementCpuWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/CpuResponse"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/CpuResponse"
        "404":
          description: "CPU not found"
          schema:
            $ref: "#/definitions/CpuResponse"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/CpuResponse"
      x-swagger-router-controller: "Cpu"
  /burnCpu:
    delete:
      tags:
      - "Cpu"
      summary: "Burn a specific CPU."
      description: "Burns a specific CPU from the world state on the blockchain.\n"
      operationId: "burnCPUDELETE"
      parameters:
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Request body with multiple variables"
        required: false
        schema:
          $ref: "#/definitions/body"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/CpuResponse"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/CpuResponse"
        "404":
          description: "CPU not found"
          schema:
            $ref: "#/definitions/CpuResponse"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/CpuResponse"
      x-swagger-router-controller: "Cpu"
  /makePayment:
    post:
      tags:
      - "Payments"
      summary: "Pay the due amount for a given bad."
      description: "Pays the due amount for the given bad in the blockchain.\n"
      operationId: "makePaymentPOST"
      parameters:
      - in: "body"
        name: "body"
        description: "payment object"
        required: true
        schema:
          $ref: "#/definitions/PaymentOperationRequestWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /updatePayment:
    post:
      tags:
      - "Payments"
      summary: "Check and update the status of a payment previously done for a given\
        \ bad"
      description: "Check and update the status of a payment previously done for a\
        \ given bad\n"
      operationId: "updatePaymenByPayNumberPOST"
      parameters:
      - in: "body"
        name: "body"
        description: "payment object"
        required: true
        schema:
          $ref: "#/definitions/statusPaymentOperationRequest"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /updatePaymentBadId:
    post:
      tags:
      - "Payments"
      summary: "Check and update the status of a payment previously done for a given\
        \ bad."
      description: "Check and update the status of a payment previously done for a\
        \ given bad\n"
      operationId: "updatePaymentByBadIdPOST"
      parameters:
      - in: "body"
        name: "body"
        description: "payment object"
        required: true
        schema:
          $ref: "#/definitions/statusPaymentOperationRequestByBadId"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /getPaymentOperationByNumber:
    get:
      tags:
      - "Payments"
      summary: "Get a payment operation by its number"
      description: "Returns the payment operation with the specified number."
      operationId: "getPaymentOperationByNumber"
      parameters:
      - name: "paymentNumber"
        in: "query"
        description: "The number of the payment operation to retrieve"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/PaymentOperation"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Payment operation not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Payments"
    x-swagger-router-controller: "Bad"
...
  /tractionExt/create:
    post:
      tags:
      - "TractionExt"
      summary: "Create a new certified traction operation"
      description: "Create a new certified traction operation"
      operationId: "createTractionExt"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "tractionExt object"
        required: true
        schema:
          $ref: "#/definitions/creationTractionExt"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Invalid tractionExt supplied"
        "401":
          description: "Authentication information is missing or invalid"
        "404":
          description: "tractionExt not found"
      x-swagger-router-controller: "TractionExt"
  /tractionExt/{TractionExtNum}:
    get:
      tags:
      - "TractionExt"
      summary: "Find tractionExt by Id"
      description: "Returns a single tractionExt identified by its TractionExtNum"
      operationId: "getTractionExtById"
      produces:
      - "application/json"
      parameters:
      - name: "TractionExtNum"
        in: "path"
        description: "ID of tractionExt to return"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/tractionExt"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "Authentication information is missing or invalid"
        "404":
          description: "tractionExt not found"
      x-swagger-router-controller: "TractionExt"
  /addPermissions:
    post:
      tags:
      - "Permissions"
      summary: "Add multiple or single permissions to a user for an organization"
      operationId: "addPermissionsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/PermissionsRequest"
      responses:
        "200":
          description: "Permissions added successfully"
      x-swagger-router-controller: "Permission"
  /updatePermissions:
    put:
      tags:
      - "Permissions"
      summary: "Update the whole permission array for a user for an organization"
      operationId: "updatePermissionsPUT"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/PermissionsRequest"
      responses:
        "200":
          description: "Permissions updated successfully"
      x-swagger-router-controller: "Permission"
  /removePermissions:
    delete:
      tags:
      - "Permissions"
      summary: "Remove multiple or single permissions from a user for an organization"
      operationId: "removePermissionsDELETE"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/PermissionsRequest"
      responses:
        "200":
          description: "Permissions removed successfully"
      x-swagger-router-controller: "Permission"
  /getPermissions:
    get:
      tags:
      - "Permissions"
      summary: "Get list of permission for a user for an organization"
      operationId: "getPermissionsGET"
      parameters:
      - name: "userName"
        in: "query"
        required: true
        type: "string"
      - name: "orgName"
        in: "query"
        required: true
        type: "string"
      responses:
        "200":
          description: "Permissions retrieved successfully"
      x-swagger-router-controller: "Permission"
  /hasPermission:
    get:
      tags:
      - "Permissions"
      summary: "Check if a user has a specific permission"
      operationId: "hasPermissionGET"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/PermissionsRequest"
      responses:
        "200":
          description: "Permission check result"
      x-swagger-router-controller: "Permission"
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  PermissionsRequest:
    type: "object"
    properties:
      userName:
        type: "string"
      orgName:
        type: "string"
      permissions:
        type: "array"
        items:
          type: "string"
    example:
      orgName: "orgName"
      permissions:
      - "permissions"
      - "permissions"
      userName: "userName"
...
  assignCpuRequest:
    type: "object"
    properties:
      badId:
        type: "string"
        description: "ID of the bad"
      nonce:
        type: "integer"
        description: "ID of the CPU"
      owner:
        type: "string"
        description: "ID of the owner"
      newOwner:
        type: "string"
        description: "ID of the new owner"
  assignCpuRequestWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/assignCpuRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
  supplementCpu:
    type: "object"
    properties:
      badId:
        type: "string"
        description: "ID of the bad"
      driverId:
        type: "string"
        description: "ID of the driver"
      nonce:
        type: "integer"
        description: "Number of the CPU"
  supplementCpuWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/supplementCpu"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
...
  signaturePayload:
    type: "string"
    description: "Signature of the body payload in base64"
    example: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
...
  Cpu:
    type: "object"
    properties:
      badId:
        type: "string"
      cpuLabel:
        type: "string"
      nounce:
        type: "integer"
      owner:
        type: "string"
      tokenValue:
        type: "string"
      minter:
        type: "string"
      tractionExt:
        type: "string"
      driver:
        type: "string"
      amp:
        type: "string"
      tractor:
        type: "string"
      truck:
        type: "string"
      plateau:
        type: "string"
    example:
      owner: "owner"
      badId: "badId"
      nounce: 0
      minter: "minter"
      driver: "driver"
      amp: "amp"
      tractor: "tractor"
      truck: "truck"
      plateau: "plateau"
      tractionExt: "tractionExt"
      cpuLabel: "cpuLabel"
      tokenValue: "tokenValue"
  CpuWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/Cpu"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        owner: "owner"
        badId: "badId"
        nounce: 0
        minter: "minter"
        driver: "driver"
        amp: "amp"
        tractor: "tractor"
        truck: "truck"
        plateau: "plateau"
        tractionExt: "tractionExt"
        cpuLabel: "cpuLabel"
        tokenValue: "tokenValue"
  PaymentOperationRequestWithSignature:
    type: "object"
    properties:
      payment:
        $ref: "#/definitions/PaymentOperationRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      payment:
        badId: "badId"
        amount: 6.02745618307040320615897144307382404804229736328125
        recipientAccountNumber: "recipientAccountNumber"
        noncePay: 0
        currency: "currency"
        senderAccountNumber: "senderAccountNumber"
  PaymentOperationRequest:
    type: "object"
    properties:
      badId:
        type: "string"
      senderAccountNumber:
        type: "string"
      recipientAccountNumber:
        type: "string"
      noncePay:
        type: "integer"
        description: "value used to generate unique ID for payment. It starts with\
          \ 0"
      amount:
        type: "number"
        description: "Integer expressing number in centimes"
      currency:
        type: "string"
    example:
      badId: "badId"
      amount: 6.02745618307040320615897144307382404804229736328125
      recipientAccountNumber: "recipientAccountNumber"
      noncePay: 0
      currency: "currency"
      senderAccountNumber: "senderAccountNumber"
  statusPaymentOperationRequest:
    type: "object"
    properties:
      paymentNumber:
        type: "string"
  statusPaymentOperationRequestByBadId:
    type: "object"
    properties:
      badId:
        type: "string"
      noncePay:
        type: "integer"
        description: "value used to generate unique ID for payment. It starts with\
          \ 0"
  PaymentOperation:
    type: "object"
    properties:
      paymentNumber:
        type: "string"
      senderAccountNumber:
        type: "string"
      recipientAccountNumber:
        type: "string"
      badId:
        type: "string"
      noncePay:
        type: "integer"
      amount:
        type: "number"
        description: "Integer expressing number in centimes"
      currency:
        type: "string"
      metadata:
        type: "string"
      status:
        type: "string"
    example:
      badId: "badId"
      amount: 6.02745618307040320615897144307382404804229736328125
      metadata: "metadata"
      recipientAccountNumber: "recipientAccountNumber"
      noncePay: 0
      currency: "currency"
      paymentNumber: "paymentNumber"
      senderAccountNumber: "senderAccountNumber"
      status: "status"
...
  CpuResponse:
    type: "object"
    properties:
      message:
        type: "string"
      error:
        type: "boolean"
    example:
      message: "message"
      error: true
...
  tractionExt:
    type: "object"
    required:
    - "declarantId"
    - "transporterId"
    properties:
      noAm:
        type: "string"
        example: "I210621447"
      cpuId:
        type: "string"
        example: "bad3-Nonce1"
      tractionExtNum:
        type: "string"
        example: "TRX15558556"
      declarantId:
        type: "string"
        example: "SAN JOSE-LOPEZ"
      transporterId:
        type: "string"
        example: "TRANS5FK"
      tractorId:
        type: "string"
        example: "12345A40"
      unitId:
        type: "string"
        example: "YDGJB537"
      cpu:
        $ref: "#/definitions/Cpu"
      source:
        type: "string"
        example: "Gate 1"
      destination:
        type: "string"
        example: "Gate 2"
      status:
        type: "string"
        description: "The status of certified pickup operation"
        default: "PENDING"
        enum:
        - "PENDING"
        - "ACCEPTED"
        - "CANCELLED"
        - "VALIDATED"
        - "REJECTED"
        - "DRAFT"
      creationTimestamp:
        type: "integer"
        format: "int32"
        description: "The timestamp of the creation of the certified pickup"
        minimum: 0
      rejectionReason:
        type: "string"
        description: "The rejection reason(by TMPA) of the certified pickup operation"
        default: "None"
    example:
      cpuId: "bad3-Nonce1"
      tractionExtNum: "TRX15558556"
      destination: "Gate 2"
      cpu:
        owner: "owner"
        badId: "badId"
        nounce: 0
        minter: "minter"
        driver: "driver"
        amp: "amp"
        tractor: "tractor"
        truck: "truck"
        plateau: "plateau"
        tractionExt: "tractionExt"
        cpuLabel: "cpuLabel"
        tokenValue: "tokenValue"
      source: "Gate 1"
      transporterId: "TRANS5FK"
      declarantId: "SAN JOSE-LOPEZ"
      creationTimestamp: 0
      tractorId: "12345A40"
      unitId: "YDGJB537"
      rejectionReason: "None"
      noAm: "I210621447"
      status: "PENDING"
  creationTractionExt:
    type: "object"
    required:
    - "declarantId"
    - "transporterId"
    properties:
      noAm:
        type: "string"
        example: "I210621447"
      cpuId:
        type: "string"
        example: "badN2-Nonce1"
      tractionExtNum:
        type: "string"
        example: "TRX15558556"
      declarantId:
        type: "string"
        example: "SAN JOSE-LOPEZ"
      transporterId:
        type: "string"
        example: "TRANS5FK"
      tractorId:
        type: "string"
        example: "12345A40"
      unitId:
        type: "string"
        example: "YDGJB537"
      source:
        type: "string"
        example: "Gate 1"
      destination:
        type: "string"
        example: "Gate 2"
      status:
        type: "string"
        description: "The status of certified pickup operation"
        default: "PENDING"
        enum:
        - "PENDING"
        - "ACCEPTED"
        - "CANCELLED"
        - "VALIDATED"
        - "REJECTED"
        - "DRAFT"
      creationTimestamp:
        type: "integer"
        format: "int32"
        description: "The timestamp of the creation of the certified pickup"
        minimum: 0
      rejectionReason:
        type: "string"
        description: "The rejection reason(by TMPA) of the certified pickup operation"
        default: "None"
...
  Message:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        minimum: 0
      message:
        type: "string"
      error:
        type: "boolean"
    example:
      code: 0
      error: true
      message: "Response Message"
...
  body:
    type: "object"
    properties:
      badId:
        type: "string"
        description: "ID of the bad"
      nonce:
        type: "integer"
        description: "ID of the CPU"